pyproject.toml
[project]
name = "autopingcode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4~=4.14.3",
    "flask~=3.1.2",
    "flask-cors~=6.0.1",
    "flask-restx~=1.3.2",
    "jsonpath~=0.82.2",
    "python-dateutil~=2.9.0.post0",
    "requests~=2.32.5",
    "urllib3~=2.6.2",
]
